pyproject.toml
[project]
name = "test-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "click==8.1.8",
    "dnspython==2.7.0",
    "email_validator==2.2.0",
    "fastapi==0.115.11",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "idna==3.10",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.6",
    "pydantic-settings==2.8.1",
    "pydantic_core==2.27.2",
    "python-dotenv==1.0.1",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.39",
    "starlette==0.46.1",
    "typing_extensions==4.12.2",
    "uvicorn==0.34.0",
]
